pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "mind"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "click>=8.1",
    "PyYAML>=6.0",
]

[project.scripts]
mind = "mind.cli:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
